requirements.txt
fastapi
uvicorn
gunicorn
requests
sqlalchemy
python-dotenv
loguru
geopy
